Fixed disappearing circles WIP
